Only allow delete scheduled or draft batch

+++ app/commands/cultivation/destroy_batch.rb
@@ -12,10 +12,15 @@ module Cultivation
     end
 
     def call
-      Cultivation::Batch.find(@batch_id).delete
-      Cultivation::Task.delete_all({batch_id: @batch_id})
-      Cultivation::TrayPlan.delete_all({batch_id: @batch_id})
-      @batch_id
+      @batch = Cultivation::Batch.find(@batch_id)
+      if @batch.status == Constants::BATCH_STATUS_DRAFT || Constants::BATCH_STATUS_SCHEDULED
+        @batch.delete
+        Cultivation::Task.delete_all({batch_id: @batch_id})
+        Cultivation::TrayPlan.delete_all({batch_id: @batch_id})
+        @batch_id
+      else
+        errors.add(:batch, "Unable to delete an active batch")
+      end
     end
   end
 end
